Listado Relacionado por ID's

+++ Dominio/_Producto.cs
@@ -14,9 +14,9 @@ namespace Dominio
         public _Categoria Categoria { get; set; }
         public _Proveedor2 Proveedor { get; set; }
         public decimal Precio { get; set; }
-        public int StockActual { get; set; }
-        public int StockMinimo { get; set; }
-        public int PorcentajeGanancia { get; set; }
+        public decimal StockActual { get; set; }
+        public decimal StockMinimo { get; set; }
+        public decimal PorcentajeGanancia { get; set; }
 
     }
 }
